+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## upcoming
+
+* added option to replace favicon
+
 ## 0.1.1
 
 * added colour picker for collections, singletons and forms settings